Organized the index.js to separate files
